+++ packages/crypto/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 0.100.1
+
+### Patch Changes
+
+- @fuel-ts/errors@0.100.1
+- @fuel-ts/utils@0.100.1
+
+## 0.100.0
+
+### Patch Changes
+
+- 6eb113c: chore: upgrade NPM package versions
+- Updated dependencies [1277ae6]
+- Updated dependencies [6eb113c]
+- Updated dependencies [f41887f]
+  - @fuel-ts/utils@0.100.0
+  - @fuel-ts/errors@0.100.0
+
 ## 0.99.0
 
 ### Patch Changes